pyproject.toml
[project]
name = "bio-benchmark"
version = "0.1.0"
description = "YAML-driven benchmark runner for bio-data-analysis agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=2.20.0",
    "huggingface-hub>=0.23.0",
    "openai>=1.55.0",
    "python-dotenv>=1.2.1",
    "PyYAML>=6.0.0",
    "requests>=2.32.0",
    "rich>=14.3.3",
]

[project.scripts]
bio-benchmark = "bio_benchmark.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
